add settings icon back, rename dashboard to routes
