+++ frontend/src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import LandingPage from './pages/Landing/Landing';
-import AdminDashboard from './pages/Admin/Dashboard';
-import RiderDashboard from './pages/Rider/Dashboard';
+import AdminRoutes from './pages/Admin/Routes';
+import RiderRoutes from './pages/Rider/Routes';
 import { AuthManager } from './components/AuthManager/AuthManager';
 import { EmployeesProvider } from './context/EmployeesContext';
 import { RidersProvider } from './context/RidersContext';
@@ -16,8 +16,8 @@ const App = () => (
         <RidersProvider>
           <Switch>
             <Route exact path="/" component={LandingPage} />
-            <Route path="/admin" component={AdminDashboard} />
-            <Route path="/rider" component={RiderDashboard} />
+            <Route path="/admin" component={AdminRoutes} />
+            <Route path="/rider" component={RiderRoutes} />
           </Switch>
         </RidersProvider>
       </EmployeesProvider>
