+++ frontend/src/icons/sidebar/index.ts
@@ -3,4 +3,5 @@ export { default as home } from './home.svg';
 export { default as riders } from './riders.svg';
 export { default as locations } from './locations.svg';
 export { default as analytics } from './analytics.svg';
+export { default as settings } from './settings.svg';
 export { default as blank } from './blank.svg';
